feat: 스케줄 생성용 시간 옵션 createTimeOptions 유틸 추가 (#67)
